+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v2.0.0
+## 10.01.2019
+1. [](#new)
+    * database agnostic with PDO from database plugin
+    * twig temaplting on process 
+    * added hardcoded values
+1. [](#fix)
+    * array field results threw exception
 # v1.0.1
 ##  08/23/2016